[Update]検索結果画面レイアウト調整(ページネーション前)
