+++ app/views/public/relationships/_follow_btn.html.erb
@@ -1,5 +1,5 @@
 <% if user.followed_by?(current_user) %>
-    <%= link_to "unfollow", user_relationships_path(user.id), class: "btn btn-block btn-success", remote: true, method: :delete %>
+    <%= link_to "フォロー中", user_relationships_path(user.id), class: "btn btn-sm btn-success", remote: true, method: :delete %>
 <% else %>
-    <%= link_to "Follow", user_relationships_path(user.id), class: "btn btn-block btn-primary", remote: true, method: :post %>
+    <%= link_to "フォローする", user_relationships_path(user.id), class: "btn btn-sm btn-primary", remote: true, method: :post %>
 <% end %>
