+++ app/views/public/searches/index.html.erb
@@ -1,20 +1,22 @@
 <div class="container col-10">
   <h1 class="row">"<%= @query %>"の検索結果</h1>
   <% if @users.present? %>
-    <h2>ユーザー</h2>
-    <div class="d-flex flex-wrap">
+    <h2 class="mt-5">ユーザー</h2>
+    <div class="d-flex justify-content-start flex-wrap ">
       <% @users.each do |user| %>
         <%= render "profile_box", user: user %>
-      <% end %>    
+      <% end %>
+      <%= paginate @users %>
     </div>
   <% end %>
   <% if @episodes.present? %>
-    <h2>エピソード</h2>
+    <h2 class="mt-5">エピソード</h2>
     <div class="d-flex flex-wrap">
       <% @episodes.each do |episode| %>
         <%= render "episode_box", episode: episode, user: episode.user %>
       <% end %>
     </div>
+    <%= paginate @episodes %>
   <% end %>
 
 </div>
